test: add multitenancy tenant model inference tests
